+++ files/default/tests/minitest/truststore_test.rb
@@ -1,0 +1,6 @@
+require File.expand_path('../support/helpers', __FILE__)
+
+describe_recipe "java-management::truststore" do
+  include Helpers::JavaManagement
+  
+end