+++ metadata.rb
@@ -4,8 +4,9 @@ maintainer_email  "bflad@wharton.upenn.edu"
 license           "Apache 2.0"
 description       "Java Management Cookbook"
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           "0.2.2"
+version           "0.3.0"
 recipe            "java-management", "Java Management"
+recipe            "java-management::truststore", "Java Truststore Management"
 
 %w{ amazon centos redhat scientific }.each do |os|
   supports os